requirements.txt
python-barcode
qrcode